feat: address changes from review
* update plugin name to Safaricom KE
* remove confusing references to Lipa na M-Pesa
* link the gateway with ngrok guide
* use asym visibility and remove trivial getters
* use libphonenumber
