+++ src/backend/app/Models/MpmPlugin.php
@@ -49,7 +49,7 @@ class MpmPlugin extends BaseModelCentral {
     public const SPARK_SHS = 28;
     public const SMS_TRANSACTION_PARSER = 29;
     public const PESAPAL_PAYMENT_PROVIDER = 30;
-    public const SAFARICOM_MOBILE_MONEY = 31;
+    public const SAFARICOM_KE_PAYMENT_PROVIDER = 31;
 
     protected $table = 'mpm_plugins';
