+++ src/backend/app/Plugins/SafaricomKePaymentProvider/Http/Requests/SafaricomSTKPushRequest.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace App\Plugins\SafaricomKePaymentProvider\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class SafaricomSTKPushRequest extends FormRequest {
+    public function authorize(): bool {
+        return true;
+    }
+
+    /**
+     * @return array<string, list<string>>
+     */
+    public function rules(): array {
+        return [
+            'amount' => ['required', 'numeric', 'min:1', 'max:150000'],
+            'phone_number' => ['required', 'string', 'min:9', 'max:15'],
+            'device_type' => ['required', 'string', 'in:meter,solar_home_system'],
+            'device_serial' => ['required', 'string', 'min:3', 'max:100'],
+            'account_reference' => ['nullable', 'string', 'max:50'],
+            'transaction_desc' => ['nullable', 'string', 'max:50'],
+            'type' => ['nullable', 'string', 'in:energy,deferred_payment,down_payment'],
+        ];
+    }
+
+    /**
+     * @return array<string, string>
+     */
+    public function messages(): array {
+        return [
+            'amount.min' => 'Amount must be at least 1 KES',
+            'amount.max' => 'Amount cannot exceed 150,000 KES per STK Push',
+        ];
+    }
+}
