+++ src/backend/app/Plugins/SafaricomKePaymentProvider/Providers/ObserverServiceProvider.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Plugins\SafaricomKePaymentProvider\Providers;
+
+use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
+
+class ObserverServiceProvider extends ServiceProvider {
+    /**
+     * Register any events for your application.
+     */
+    public function boot(): void {
+        parent::boot();
+    }
+}
