--- src/backend/app/Plugins/SafaricomMobileMoney/Providers/ObserverServiceProvider.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace App\Plugins\SafaricomMobileMoney\Providers;
-
-use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
-
-class ObserverServiceProvider extends ServiceProvider {
-    /**
-     * Register any events for your application.
-     */
-    public function boot(): void {
-        parent::boot();
-    }
-}
